+++ ccramic/utils/alert.py
@@ -67,7 +67,12 @@ class AlertMessage(BaseModel):
                                                      " (typically 2x the size of the imported files).",
                       "lazy-load-error": "Error when loading data from the imported file. Check that the dataset "
                                          "delimiter does not have any overlapping characters with any of the filenames, "
-                                         "or ROI names. "}
+                                         "or ROI names. ",
+                      "invalid_query": "Error when querying from quantification results. Ensure that: \n"
+                                       "\n-quantification results are loaded"
+                                       "\n-the corresponding images for quantified ROIs are loaded, and an image "
+                                       "has been generated in the main canvas"
+                                       "\n-ROI naming in the quantification sheet matches the ROI names in the session."}
 
 
 class ToolTips(BaseModel):